Cleanup utf8

+++ src/utf8.c
@@ -239,9 +239,9 @@ int xgetc_utf8(int(*fn)(), void *p1)
 	return (int)n;
 }
 
-int index_utf8(const char *src, short i)
+int index_utf8(const char *src, size_t i)
 {
-	short idx = 0;
+	size_t idx = 0;
 
 	while (*src) {
 		int ch = get_char_utf8(&src);
