+++ src/utf8.h
@@ -40,4 +40,4 @@ extern size_t len_char_utf8(const char *src);					// returns #bytes
  *  Get indexed char
  */
 
-extern int index_utf8(const char *src, short i);
+extern int index_utf8(const char *src, size_t i);
